@@ -1,10 +1,10 @@
 #!/usr/bin/env kotlin
 
-@file:DependsOn("io.github.typesafegithub:github-workflows-kt:1.6.0")
+@file:DependsOn("io.github.typesafegithub:github-workflows-kt:1.9.0")
 
 
 import io.github.typesafegithub.workflows.actions.actions.CheckoutV4
-import io.github.typesafegithub.workflows.actions.actions.SetupJavaV3
+import io.github.typesafegithub.workflows.actions.actions.SetupJavaV4
 import io.github.typesafegithub.workflows.domain.RunnerType
 import io.github.typesafegithub.workflows.domain.Workflow
 import io.github.typesafegithub.workflows.domain.actions.CustomAction
@@ -16,7 +16,7 @@ import java.nio.file.Paths
 public val workflowFormat: Workflow = workflow(
     name = "Format",
     on = listOf(Push()),
-    sourceFile = Paths.get(".github/kts/format.main.kts"),
+    sourceFile =  __FILE__.toPath(),
 ) {
     job(
         id = "formatting",
@@ -30,8 +30,8 @@ public val workflowFormat: Workflow = workflow(
         )
         uses(
             name = "JDK 17",
-            action = SetupJavaV3(
-                distribution = SetupJavaV3.Distribution.Custom("corretto"),
+            action = SetupJavaV4(
+                distribution = SetupJavaV4.Distribution.Corretto,
                 javaVersion = "17",
             ),
         )
@@ -44,7 +44,7 @@ public val workflowFormat: Workflow = workflow(
             action = CustomAction(
                 actionOwner = "stefanzweifel",
                 actionName = "git-auto-commit-action",
-                actionVersion = "v4",
+                actionVersion = "v5",
                 inputs = mapOf(
                     "commit_message" to "Apply formatter",
                 )
