@@ -27,9 +27,10 @@ import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
 import java.io.IOException;
 import java.time.Duration;
-import java.util.*;
-import java.util.concurrent.*;
-import java.util.function.*;
+import java.util.Timer;
+import java.util.TimerTask;
+import java.util.concurrent.TimeoutException;
+import java.util.function.Consumer;
 import lombok.NonNull;
 import lombok.SneakyThrows;
 import org.factcast.client.grpc.FactCastGrpcClientProperties.ResilienceConfiguration;
@@ -37,6 +38,8 @@ import org.factcast.client.grpc.ResilientGrpcSubscription.DelegatingFactObserver
 import org.factcast.client.grpc.ResilientGrpcSubscription.SubscriptionHolder;
 import org.factcast.client.grpc.ResilientGrpcSubscription.ThrowingBiConsumer;
 import org.factcast.core.Fact;
+import org.factcast.core.FactStreamPosition;
+import org.factcast.core.TestFactStreamPosition;
 import org.factcast.core.store.RetryableException;
 import org.factcast.core.subscription.FactStreamInfo;
 import org.factcast.core.subscription.Subscription;
@@ -293,9 +296,9 @@ class ResilientGrpcSubscriptionTest {
 
     @Test
     void ffwDelegates() {
-      @NonNull UUID id = UUID.randomUUID();
-      dfo.onFastForward(id);
-      verify(obs).onFastForward(id);
+      FactStreamPosition pos = TestFactStreamPosition.random();
+      dfo.onFastForward(pos);
+      verify(obs).onFastForward(pos);
     }
 
     @Test
