@@ -93,9 +93,15 @@ public class DefaultFact implements Fact, Externalizable {
       throw new IllegalArgumentException("version attribute is not valid " + jsonHeader);
   }
 
+  /**
+   * @deprecated use header.meta(String) instead
+   * @param key
+   * @return null if does not exist
+   */
   @Override
-  public String meta(String key) {
-    return deserializedHeader.meta().get(key);
+  @Deprecated
+  public String meta(@NonNull String key) {
+    return deserializedHeader.meta(key);
   }
 
   @Override
@@ -115,7 +121,7 @@ public class DefaultFact implements Fact, Externalizable {
   }
 
   @Override
-  public UUID id() {
+  public @NonNull UUID id() {
     return deserializedHeader.id();
   }
 
@@ -125,7 +131,7 @@ public class DefaultFact implements Fact, Externalizable {
   }
 
   @Override
-  public String ns() {
+  public @NonNull String ns() {
     return deserializedHeader.ns();
   }
 
