@@ -16,9 +16,10 @@
 package org.factcast.factus.projector;
 
 import java.util.List;
-import java.util.UUID;
+import javax.annotation.Nullable;
 import lombok.NonNull;
 import org.factcast.core.Fact;
+import org.factcast.core.FactStreamPosition;
 import org.factcast.core.spec.FactSpec;
 import org.factcast.factus.projection.Projection;
 
@@ -31,5 +32,5 @@ public interface Projector<A extends Projection> {
 
   List<FactSpec> createFactSpecs();
 
-  void onCatchup(UUID idOfLastFactApplied);
+  void onCatchup(@Nullable FactStreamPosition lastFactApplied);
 }
