@@ -16,7 +16,7 @@
 package org.factcast.factus.projector;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.mock;
 
 import java.lang.reflect.Method;
 import java.util.Collections;
@@ -29,16 +29,18 @@ import lombok.Value;
 import org.assertj.core.util.Maps;
 import org.factcast.core.Fact;
 import org.factcast.core.FactHeader;
+import org.factcast.core.FactStreamPosition;
+import org.factcast.core.TestFactStreamPosition;
 import org.factcast.core.event.EventConverter;
 import org.factcast.core.spec.FactSpec;
 import org.factcast.core.util.FactCastJson;
 import org.factcast.factus.*;
 import org.factcast.factus.event.DefaultEventSerializer;
 import org.factcast.factus.event.EventSerializer;
-import org.factcast.factus.projection.LocalManagedProjection;
 import org.factcast.factus.projection.Projection;
 import org.factcast.factus.projector.ProjectorImpl.ReflectionTools;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
 
 /** Blackbox test, we are wiring real objects into the test class, no mocks. */
 class ProjectorImplTest {
@@ -183,7 +185,8 @@ class ProjectorImplTest {
     @Test
     void createSimple() {
       // INIT
-      UUID factStreamPosition = UUID.fromString("9258562c-e6aa-4855-a765-3b1f49a113d5");
+      FactStreamPosition factStreamPosition =
+          TestFactStreamPosition.fromString("9258562c-e6aa-4855-a765-3b1f49a113d5");
 
       ComplexProjection projection = new ComplexProjection();
       projection.factStreamPosition(factStreamPosition);
@@ -389,45 +392,6 @@ class ProjectorImplTest {
       Method realMethod =
           NonStaticClass$MockitoMock.class.getDeclaredMethod("apply", SimpleEvent.class);
       assertThat(underTest.isEventHandlerMethod(realMethod)).isFalse();
-    }
-  }
-
-  @Nested
-  class WhenTestingCatchingUp {
-
-    @Test
-    void setsFactStreamPosition() throws NoSuchMethodException {
-
-      LocalManagedProjection factStreamPositionAware =
-          spy(
-              new LocalManagedProjection() {
-                @Handler
-                void apply(SimpleEvent e) {}
-              });
-      ProjectorImpl<LocalManagedProjection> underTest =
-          new ProjectorImpl<>(mock(EventSerializer.class), factStreamPositionAware);
-
-      UUID id = UUID.randomUUID();
-      underTest.onCatchup(id);
-
-      verify(factStreamPositionAware).factStreamPosition(id);
-    }
-
-    @Test
-    void skipsSettingFactStreamPosition() throws NoSuchMethodException {
-      Projection notFactStreamPositionAware =
-          spy(
-              new Projection() {
-                @Handler
-                void apply(SimpleEvent e) {}
-              });
-      ProjectorImpl<Projection> underTest =
-          new ProjectorImpl<>(mock(EventSerializer.class), notFactStreamPositionAware);
-
-      UUID id = UUID.randomUUID();
-      underTest.onCatchup(id);
-
-      verifyNoInteractions(notFactStreamPositionAware);
     }
   }
 
