@@ -18,11 +18,13 @@ package org.factcast.grpc.api.conv;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.ProtocolStringList;
+import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.factcast.core.Fact;
+import org.factcast.core.FactStreamPosition;
 import org.factcast.core.snap.Snapshot;
 import org.factcast.core.snap.SnapshotId;
 import org.factcast.core.spec.FactSpec;
@@ -81,10 +83,11 @@ public class ProtoConverter {
   }
 
   @NonNull
-  public MSG_Notification createNotificationForFastForward(@NonNull UUID id) {
+  public MSG_Notification createNotificationForFastForward(@NonNull FactStreamPosition id) {
     MSG_Notification.Builder builder =
         MSG_Notification.newBuilder().setType(MSG_Notification.Type.Ffwd);
-    builder.setId(toProto(id));
+    builder.setId(toProto(id.factId()));
+    builder.setSerial(toProto(id.serial()));
     return builder.build();
   }
 
@@ -316,7 +319,7 @@ public class ProtoConverter {
   }
 
   @NonNull
-  public MSG_CurrentDatabaseTime toProto(long currentTime) {
+  public MSG_CurrentDatabaseTime toProtoTime(long currentTime) {
     return MSG_CurrentDatabaseTime.newBuilder().setMillis(currentTime).build();
   }
 
@@ -418,5 +421,29 @@ public class ProtoConverter {
 
   public MSG_Notification createInfoNotification(FactStreamInfo info) {
     return MSG_Notification.newBuilder().setType(Type.Info).setInfo(toProto(info)).build();
+  }
+
+  public long fromProto(MSG_Serial msgSerial) {
+    return msgSerial.getSerial();
+  }
+
+  public MSG_Date toProto(LocalDate date) {
+    return MSG_Date.newBuilder()
+        .setYear(date.getYear())
+        .setMonth(date.getMonthValue())
+        .setDay(date.getDayOfMonth())
+        .build();
+  }
+
+  public LocalDate fromProto(MSG_Date msgDate) {
+    return LocalDate.of(msgDate.getYear(), msgDate.getMonth(), msgDate.getDay());
+  }
+
+  public MSG_Serial toProto(long serial) {
+    return MSG_Serial.newBuilder().setSerial(serial).build();
+  }
+
+  public FactStreamPosition fromProto(MSG_UUID id, MSG_Serial serial) {
+    return FactStreamPosition.of(fromProto(id), fromProto(serial));
   }
 }
