@@ -15,15 +15,19 @@
  */
 package org.factcast.itests.factus.client;
 
-import static java.util.UUID.*;
+import static java.util.UUID.randomUUID;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.Duration;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.UUID;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.factcast.core.FactStreamPosition;
 import org.factcast.factus.Factus;
 import org.factcast.factus.Handler;
 import org.factcast.factus.event.EventObject;
@@ -446,12 +450,13 @@ public class SpringTransactionalITest extends AbstractFactCastIntegrationTest {
     }
 
     @Override
-    public UUID factStreamPosition() {
+    public FactStreamPosition factStreamPosition() {
       try {
-        return jdbcTemplate.queryForObject(
-            "SELECT fact_stream_position FROM managed_projection WHERE name = ?",
-            UUID.class,
-            getScopedName().asString());
+        return FactStreamPosition.withoutSerial(
+            jdbcTemplate.queryForObject(
+                "SELECT fact_stream_position FROM managed_projection WHERE name = ?",
+                UUID.class,
+                getScopedName().asString()));
       } catch (IncorrectResultSizeDataAccessException e) {
         // no position yet, just return null
         return null;
@@ -459,7 +464,7 @@ public class SpringTransactionalITest extends AbstractFactCastIntegrationTest {
     }
 
     @Override
-    public void factStreamPosition(@NonNull UUID factStreamPosition) {
+    public void factStreamPosition(@NonNull FactStreamPosition factStreamPosition) {
       log.debug("set fact stream position");
       assertThat(TransactionSynchronizationManager.isActualTransactionActive()).isTrue();
       factStreamPositionModifications++;
@@ -470,8 +475,8 @@ public class SpringTransactionalITest extends AbstractFactCastIntegrationTest {
           "INSERT INTO managed_projection (name, fact_stream_position) VALUES (?, ?) ON CONFLICT"
               + " (name) DO UPDATE SET fact_stream_position = ?",
           getScopedName().asString(),
-          factStreamPosition,
-          factStreamPosition);
+          factStreamPosition.factId(),
+          factStreamPosition.factId());
     }
 
     @Override
@@ -503,12 +508,13 @@ public class SpringTransactionalITest extends AbstractFactCastIntegrationTest {
     }
 
     @Override
-    public UUID factStreamPosition() {
+    public FactStreamPosition factStreamPosition() {
       try {
-        return jdbcTemplate.queryForObject(
-            "SELECT fact_stream_position FROM managed_projection WHERE name = ?",
-            UUID.class,
-            getScopedName().asString());
+        return FactStreamPosition.withoutSerial(
+            jdbcTemplate.queryForObject(
+                "SELECT fact_stream_position FROM managed_projection WHERE name = ?",
+                UUID.class,
+                getScopedName().asString()));
       } catch (IncorrectResultSizeDataAccessException e) {
         // no position yet, just return null
         return null;
@@ -516,7 +522,7 @@ public class SpringTransactionalITest extends AbstractFactCastIntegrationTest {
     }
 
     @Override
-    public void factStreamPosition(@NonNull UUID state) {
+    public void factStreamPosition(@NonNull FactStreamPosition state) {
       log.debug("set state");
       assertThat(TransactionSynchronizationManager.isActualTransactionActive()).isTrue();
       factStreamPositionModifications++;
@@ -527,8 +533,8 @@ public class SpringTransactionalITest extends AbstractFactCastIntegrationTest {
           "INSERT INTO managed_projection (name, fact_stream_position) VALUES (?, ?) "
               + "ON CONFLICT (name) DO UPDATE SET fact_stream_position = ?",
           getScopedName().asString(),
-          state,
-          state);
+          state.factId(),
+          state.factId());
     }
 
     @Override
