@@ -17,13 +17,16 @@ package org.factcast.store.internal;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.eventbus.EventBus;
-import java.util.*;
-import java.util.concurrent.atomic.*;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicLong;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.factcast.core.FactStreamPosition;
 import org.factcast.core.subscription.FactStreamInfo;
 import org.factcast.core.subscription.SubscriptionImpl;
 import org.factcast.core.subscription.SubscriptionRequest;
@@ -186,7 +189,7 @@ public class PgFactStream {
       long targetSer = ffwdTarget.targetSer();
 
       if (targetId != null && (targetSer > startedSer)) {
-        subscription.notifyFastForward(targetId);
+        subscription.notifyFastForward(FactStreamPosition.of(targetId, targetSer));
       }
     }
   }
