@@ -16,15 +16,12 @@
 package org.factcast.store.registry.filesystem;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.util.Optional;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.factcast.store.registry.IndexFetcher;
 import org.factcast.store.registry.RegistryIndex;
-import org.factcast.store.registry.SchemaRegistryUnavailableException;
-import org.factcast.store.registry.http.ValidationConstants;
+import org.springframework.core.io.FileSystemResource;
 
 @RequiredArgsConstructor
 public class FilesystemIndexFetcher implements IndexFetcher {
@@ -34,20 +31,6 @@ public class FilesystemIndexFetcher implements IndexFetcher {
   @Override
   public Optional<RegistryIndex> fetchIndex() {
     File file = new File(base, "index.json");
-    return fetch_index(file);
-  }
-
-  @NonNull
-  public static Optional<RegistryIndex> fetch_index(File file) {
-    try {
-      if (file.exists()) {
-        return Optional.of(ValidationConstants.JACKSON.readValue(file, RegistryIndex.class));
-      } else {
-        throw new SchemaRegistryUnavailableException(
-            new FileNotFoundException("Resource " + file + " does not exist."));
-      }
-    } catch (IOException e) {
-      throw new SchemaRegistryUnavailableException(e);
-    }
+    return RegistryIndex.fetch(new FileSystemResource(file));
   }
 }
