@@ -15,9 +15,10 @@
  */
 package org.factcast.store.internal;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.Mockito.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.same;
+import static org.mockito.Mockito.verify;
 
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
@@ -34,10 +35,14 @@ import lombok.SneakyThrows;
 import lombok.experimental.Delegate;
 import org.assertj.core.util.Lists;
 import org.factcast.core.Fact;
+import org.factcast.core.FactStreamPosition;
 import org.factcast.core.snap.Snapshot;
 import org.factcast.core.snap.SnapshotId;
 import org.factcast.core.spec.FactSpec;
-import org.factcast.core.store.*;
+import org.factcast.core.store.FactStore;
+import org.factcast.core.store.State;
+import org.factcast.core.store.StateToken;
+import org.factcast.core.store.TokenStore;
 import org.factcast.core.subscription.SubscriptionRequest;
 import org.factcast.core.subscription.SubscriptionRequestTO;
 import org.factcast.core.subscription.observer.FactObserver;
@@ -156,7 +161,7 @@ class PgFactStoreIntegrationTest extends AbstractFactStoreTest {
     @NonNull final UUID id = UUID.randomUUID();
     @NonNull final UUID id2 = UUID.randomUUID();
     @NonNull final UUID id3 = UUID.randomUUID();
-    final AtomicReference<UUID> fwd = new AtomicReference<>();
+    final AtomicReference<FactStreamPosition> fwd = new AtomicReference<>();
 
     @NonNull
     final FactObserver obs =
@@ -171,7 +176,7 @@ class PgFactStoreIntegrationTest extends AbstractFactStoreTest {
           }
 
           @Override
-          public void onFastForward(@NonNull UUID factIdToFfwdTo) {
+          public void onFastForward(@NonNull FactStreamPosition factIdToFfwdTo) {
             fwd.set(factIdToFfwdTo);
             System.out.println("ffwd " + factIdToFfwdTo);
           }
@@ -242,7 +247,7 @@ class PgFactStoreIntegrationTest extends AbstractFactStoreTest {
 
       // ffwd expected
       assertThat(fwd.get()).isNotNull();
-      UUID first = fwd.get();
+      FactStreamPosition first = fwd.get();
 
       // publish unrelated stuff and update ffwd target
       store.publish(
@@ -278,8 +283,8 @@ class PgFactStoreIntegrationTest extends AbstractFactStoreTest {
   }
 
   @Nested
-  class LocalFactStoreTest {
-    final LocalFactStore localFactStore = (LocalFactStore) fs;
+  class FactStoreTest {
+    final FactStore localFactStore = fs;
 
     @Test
     void testFetchBySerialReturnsFact() {
